dataStructures/pkg/arrays/challenges/trap_rainwater.go
package challenges

func trap(height []int) int {
	n := len(height)
	if n < 3 {
		return 0
	}

	leftMax := make([]int, n)
	rightMax := make([]int, n)

	leftMax[0] = height[0]
	for i := 1; i < n; i++ {
		leftMax[i] = max(leftMax[i-1], height[i])
	}

	rightMax[n-1] = height[n-1]
	for i := n - 2; i >= 0; i-- {
		rightMax[i] = max(rightMax[i+1], height[i])
	}

	trapped := 0
	for i := 0; i < n; i++ {
		trapped += min(leftMax[i], rightMax[i]) - height[i]
	}

	return trapped
}

func max(a, b int) int {
	if a > b {
		return a
	}
	return b
}

func min(a, b int) int {
	if a < b {
		return a
	}
	return b
}
